+++ src/Parser/AbstractFileReader.php
@@ -36,40 +36,4 @@ abstract class AbstractFileReader implements ParserInterface
 
         return $this->parse($contents);
     }
-
-    /**
-     * @deprecated since 2.0.0, use parse() instead. Will be removed in a
-     *             future version.
-     */
-    public function fromString(string $string): array
-    {
-        @trigger_error(
-            sprintf(
-                '%s::fromString() is deprecated since 2.0.0, use %s::parse() instead. It will be removed in a future version.',
-                static::class,
-                static::class
-            ),
-            E_USER_DEPRECATED
-        );
-
-        return $this->parse($string);
-    }
-
-    /**
-     * @deprecated since 2.0.0, use parseFile() instead. Will be removed in a
-     *             future version.
-     */
-    public function fromFile(string $fileName): array
-    {
-        @trigger_error(
-            sprintf(
-                '%s::fromFile() is deprecated since 2.0.0, use %s::parseFile() instead. It will be removed in a future version.',
-                static::class,
-                static::class
-            ),
-            E_USER_DEPRECATED
-        );
-
-        return $this->parseFile($fileName);
-    }
 }